cmake: fix undefined symbols and multiple definitions
