+++ src/device/CMakeLists.txt
@@ -71,6 +71,7 @@ target_link_libraries(device
   PUBLIC
     ${HIDAPI_LIBRARIES}
     cncrypto
+    cryptonote_format_utils_basic
     ringct_basic
     wallet-crypto
     ${OPENSSL_CRYPTO_LIBRARIES}
